requirements.txt
APScheduler==3.10.4
requests==2.31.0
PyYAML==6.0.1
pytz==2024.1
